Issue #258 - Add validate command and strict unmarshalling
Required an update to the yaml package in order to use
unmarshalStrict.

Validate method currently runs and returns null. Based on the
issue description it seemed appropriate to always do a strict
check on the config and let the errors thrown there take care
of actual validation. The validate command is really just
a null method that loads the config like normal to perform
checks.