Issue #273 - Added default policy and changed upsert signature
A stack policy exists in the assets and is currently added to
all stacks. The policy does not allow replace or delete actions
on RDS or KMS resources.

I plan to pass the policy in as a string to upsert stack as
a way to set the policy later through extensions or cli flags.