Elb: if there's a cert, redirect http to https
Currently if a user visits a mu elb with http, they are not redirected
to https. In the mu spirit of taking opinionated stances, the right
default behavior should be to redirect to https where available.

The cloudformation support for redirect didn't exist a month ago;
I suspect it was quietly released around "re: invent".